style: refactor package
