+++ src/main/java/chess/console/consoleview/boardview/BoardView.java
@@ -1,4 +1,4 @@
-package chess.view.boardview;
+package chess.console.consoleview.boardview;
 
 import static chess.domain.position.File.FILES_TOTAL_SIZE;
 import static chess.domain.position.Rank.RANKS_TOTAL_SIZE;
