+++ src/main/java/chess/web/util/JsonTransformer.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.web.util;
 
 import com.google.gson.Gson;
 import spark.ResponseTransformer;
